+++ controller/productCon.js
@@ -13,14 +13,29 @@ export const getProducts = async (req, res) => {
 export const getProductById = async (req, res) => {
     try {
         const { _id } = req.params;
-        const product = await Product.findOne({_id: _id});
-        res.status(200).json(product);
+        const product = await Product.findById(_id);
+        if(product === null){
+            res.status(404).json();
+        }
+        if(product){
+            res.status(200).json(product);
+        }
     } catch (err) {
         res.status(500).json({message: err});
     }
 }
 
 export const createProduct = async (req, res) => {
+    try {
+        const newProducts = req.body;
+        const products = await Product.create(newProducts);
+        res.status(200).json(products);
+    } catch (err) {
+        res.status(422).json({message: 'Validation Error'});
+    }
+}
+
+export const createManyProduct = async (req, res) => {
     try {
         const newProducts = req.body;
         const products = await Product.insertMany(newProducts);
@@ -33,8 +48,15 @@ export const createProduct = async (req, res) => {
 export const updateProduct = async (req, res) => {
     try {
         const updateProduct = req.body;
-        const product = await Product.findByIdAndUpdate(updateProduct._id, updateProduct, {new: true});
-        res.status(200).json(product);
+        const handle = typeof req.body._id === 'number' && typeof req.body.prodName === 'string' && typeof req.body.prodNumber === 'number' && typeof req.body.prodPrice === 'number' && typeof req.body.prodSale === 'number' && typeof req.body.prodImageUrl === 'string' && typeof req.body.prodCategories === 'object' && typeof req.body.prodAuthor === 'object';
+
+        if(!handle){
+            res.status(422).json()
+        }
+        if(handle){
+            const product = await Product.findByIdAndUpdate(updateProduct._id, updateProduct, {new: true});
+            res.status(200).json(product)
+        }
     } catch (err) {
         res.status(500).json({message: err});
     }
@@ -43,8 +65,15 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     try {
         const { _id } = req.params;
-        const product = await Product.findByIdAndDelete(_id);
-        res.status(204).json(product);
+        const product = await Product.findById(_id);
+
+        if(product === null){
+            res.status(404).json();
+        }
+        if(product){
+            const product = await Product.findByIdAndDelete(_id);
+            res.status(204).json(product);
+        }
     } catch (err) {
         res.status(500).json({message: err});
     }
